+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/csvtk/v0.30.1/total.svg)](https://github.com/shenwei356/csvtk/releases/tag/v0.30.1)
     - `csvtk sort/join`:
         - faster speed and lower memory.
+    - `csvtk summary`:
+        - fix the order of columns. [#282](https://github.com/shenwei356/csvtk/issues/282)
     - `csvtk spread`:
         - add a new alias "scatter" to "spread". [#265](https://github.com/shenwei356/csvtk/issues/265)
     - `csvtk grep`: